requirements.txt
# 데이터 수집 (한국 주식, 무료, API 키 불필요)
finance-datareader>=0.9.50
pandas>=2.0.0
numpy>=1.24.0

# 기술적 지표
ta>=0.10.2

# 환경 변수
python-dotenv>=1.0.0

# 대시보드 (추천 결과 확인)
streamlit>=1.28.0

# 모바일 웹앱 (조회 버튼)
flask>=3.0.0
gunicorn>=21.0.0

# AI 분석 (선택)
# openai>=1.0.0
